move GOG/innoextract import out of useSettings into useGogImport
Half of useSettings was importer logic (innoextract discovery, installer
extraction, WAD collection) that only reads the library path. It now
lives in useGogImport, which also resolves iwads/ through useLibrary
instead of joining the path by hand, and documents that the command
names must stay in sync with the shell allowlist in capabilities.
